Implement a simpler preferences window
- Fix #29
- Keep the SHIFT down to open the old "Advanced Preferences Window"
