+++ Source/Makefile
@@ -28,7 +28,8 @@ APP_MIME_SIG =
 #	means this Makefile will not work correctly if two source files with the
 #	same name (source.c or source.cpp) are included from different directories.
 #	Also note that spaces in folder names do not work well with this Makefile.
-SRCS = CheckBoxWithStringColumn.cpp \
+SRCS = AdvancedPrefsWindow.cpp \
+	CheckBoxWithStringColumn.cpp \
 	ElementListView.cpp \
 	EraserLooper.cpp \
 	FilWip.cpp \
